Separates tests that need the platform running to extend integration test base class.
